+++ adapters/Facebook/FacebookAdapter/GADFBNativeAd.m
@@ -23,7 +23,6 @@
 #import "GADFBNetworkExtras.h"
 
 static NSString *const GADNativeAdIconView = @"2003";
-static NSString *const GADUnifiedNativeAdIconView = @"3003";
 
 @interface GADFBNativeAd () <GADMediatedNativeAppInstallAd,
                              GADMediatedNativeAdDelegate,
@@ -51,8 +50,8 @@ static NSString *const GADUnifiedNativeAdIconView = @"3003";
   /// Serializes ivar usage.
   dispatch_queue_t _lockQueue;
 
-  /// Facebook AdChoices view.
-  FBAdChoicesView *_adChoicesView;
+  /// Facebook AdOptions view.
+  FBAdOptionsView *_adOptionsView;
 
   /// YES if an impression has been logged.
   BOOL _impressionLogged;
@@ -135,20 +134,31 @@ static NSString *const GADUnifiedNativeAdIconView = @"3003";
   _mediaView.delegate = nil;
 }
 
-- (void)loadAdChoicesView {
-  id<GADMAdNetworkConnector> strongConnector = _connector;
-  id obj = [strongConnector networkExtras];
-  GADFBNetworkExtras *networkExtras = [obj isKindOfClass:[GADFBNetworkExtras class]] ? obj : nil;
-  if (!_adChoicesView) {
-    if (networkExtras) {
-      _adChoicesView = [[FBAdChoicesView alloc] initWithNativeAd:self->_nativeAd
-                                                      expandable:networkExtras.adChoicesExpandable];
-      _adChoicesView.backgroundShown = networkExtras.adChoicesBackgroundShown;
-    } else {
-      _adChoicesView = [[FBAdChoicesView alloc] initWithNativeAd:self->_nativeAd];
-    }
+- (void)loadAdOptionsView {
+  if (!_adOptionsView) {
+    _adOptionsView = [[FBAdOptionsView alloc] init];
+    _adOptionsView.backgroundColor = [UIColor clearColor];
+
+    NSLayoutConstraint *height =
+        [NSLayoutConstraint constraintWithItem:_adOptionsView
+                                     attribute:NSLayoutAttributeHeight
+                                     relatedBy:NSLayoutRelationEqual
+                                        toItem:nil
+                                     attribute:NSLayoutAttributeNotAnAttribute
+                                    multiplier:0
+                                      constant:FBAdOptionsViewHeight];
+    NSLayoutConstraint *width =
+        [NSLayoutConstraint constraintWithItem:_adOptionsView
+                                     attribute:NSLayoutAttributeWidth
+                                     relatedBy:NSLayoutRelationEqual
+                                        toItem:nil
+                                     attribute:NSLayoutAttributeNotAnAttribute
+                                    multiplier:0
+                                      constant:FBAdOptionsViewWidth];
+    [_adOptionsView addConstraint:height];
+    [_adOptionsView addConstraint:width];
+    [_adOptionsView updateConstraints];
   }
-  [_adChoicesView updateFrameFromSuperview];
 }
 
 #pragma mark - GADMediatedNativeAd
@@ -235,7 +245,7 @@ static NSString *const GADUnifiedNativeAdIconView = @"3003";
 }
 
 - (UIView *GAD_NULLABLE_TYPE)adChoicesView {
-  return _adChoicesView;
+  return _adOptionsView;
 }
 
 /// Returns YES if the ad has video content.
@@ -257,8 +267,6 @@ static NSString *const GADUnifiedNativeAdIconView = @"3003";
 
   if ([view isKindOfClass:[GADNativeAppInstallAdView class]]) {
     iconView = [clickableAssetViews valueForKey:GADNativeAdIconView];
-  } else if ([view isKindOfClass:[GADUnifiedNativeAdView class]]) {
-    iconView = [clickableAssetViews valueForKey:GADUnifiedNativeAdIconView];
   }
 
   if (assets.count > 0 && iconView) {
@@ -276,7 +284,6 @@ static NSString *const GADUnifiedNativeAdIconView = @"3003";
 }
 
 - (void)mediatedNativeAd:(id<GADMediatedNativeAd>)mediatedNativeAd didUntrackView:(UIView *)view {
-  [_adChoicesView removeFromSuperview];
   [_nativeAd unregisterView];
 }
 
@@ -285,7 +292,7 @@ static NSString *const GADUnifiedNativeAdIconView = @"3003";
 - (void)nativeAdDidLoad:(FBNativeAd *)nativeAd {
   _mediaView = [[FBMediaView alloc] init];
   _mediaView.delegate = self;
-  [self loadAdChoicesView];
+  [self loadAdOptionsView];
   id<GADMAdNetworkAdapter> strongAdapter = self->_adapter;
   id<GADMAdNetworkConnector> strongConnector = self->_connector;
   [strongConnector adapter:strongAdapter didReceiveMediatedNativeAd:self];