+++ adapters/Unity/UnityAdapter/GADMAdapterUnity.m
@@ -16,14 +16,12 @@
 
 #import "GADMAdapterUnityConstants.h"
 #import "GADMAdapterUnitySingleton.h"
+#import "GADMediationAdapterUnity.h"
 #import "GADUnityError.h"
 
 @interface GADMAdapterUnity () {
   /// Connector from Google Mobile Ads SDK to receive ad configurations.
-  __weak id<GADMRewardBasedVideoAdNetworkConnector> _rewardBasedVideoAdConnector;
-
-  /// Connector from Google Mobile Ads SDK to receive ad configurations.
-  __weak id<GADMAdNetworkConnector> _interstitialConnector;
+  __weak id<GADMAdNetworkConnector> _networkConnector;
 
   /// Placement ID of Unity Ads network.
   NSString *_placementID;
@@ -36,68 +34,16 @@
 
 @implementation GADMAdapterUnity
 
++ (nonnull Class<GADMediationAdapter>)mainAdapterClass {
+  return [GADMediationAdapterUnity class];
+}
+
 + (NSString *)adapterVersion {
-  return GADMAdapterUnityVersion;
+  return kGADMAdapterUnityVersion;
 }
 
 + (Class<GADAdNetworkExtras>)networkExtrasClass {
   return Nil;
-}
-
-#pragma mark Reward-based Video Ad Methods
-
-- (instancetype)initWithRewardBasedVideoAdNetworkConnector:
-        (id<GADMRewardBasedVideoAdNetworkConnector>)connector {
-  if (!connector) {
-    return nil;
-  }
-
-  self = [super init];
-  if (self) {
-    _rewardBasedVideoAdConnector = connector;
-  }
-  return self;
-}
-
-- (void)setUp {
-  id<GADMRewardBasedVideoAdNetworkConnector> strongConnector = _rewardBasedVideoAdConnector;
-  NSString *gameID =
-      [[[strongConnector credentials] objectForKey:GADMAdapterUnityGameID] copy];
-  _placementID =
-      [[[strongConnector credentials] objectForKey:GADMAdapterUnityPlacementID] copy];
-  if (!gameID || !_placementID) {
-    NSError *error = GADUnityErrorWithDescription(@"Game ID and Placement ID cannot be nil.");
-    [strongConnector adapter:self didFailToSetUpRewardBasedVideoAdWithError:error];
-    return;
-  }
-  BOOL isConfigured =
-      [[GADMAdapterUnitySingleton sharedInstance] configureRewardBasedVideoAdWithGameID:gameID
-                                                                               delegate:self];
-  if (isConfigured) {
-    [strongConnector adapterDidSetUpRewardBasedVideoAd:self];
-  } else {
-    NSString *description =
-        [[NSString alloc] initWithFormat:@"%@ is not supported for this device.",
-                                         NSStringFromClass([UnityAds class])];
-    NSError *error = GADUnityErrorWithDescription(description);
-    [strongConnector adapter:self didFailToSetUpRewardBasedVideoAdWithError:error];
-  }
-}
-
-- (void)requestRewardBasedVideoAd {
-  _placementID =
-      [[[_rewardBasedVideoAdConnector credentials] objectForKey:GADMAdapterUnityPlacementID] copy];
-  _isLoading = YES;
-  [[GADMAdapterUnitySingleton sharedInstance] requestRewardBasedVideoAdWithDelegate:self];
-}
-
-- (void)presentRewardBasedVideoAdWithRootViewController:(UIViewController *)viewController {
-  // We will send adapterDidOpenRewardBasedVideoAd callback before presenting the unity ad because
-  // the ad has already loaded.
-  [_rewardBasedVideoAdConnector adapterDidOpenRewardBasedVideoAd:self];
-  [[GADMAdapterUnitySingleton sharedInstance]
-      presentRewardBasedVideoAdForViewController:viewController
-                                        delegate:self];
 }
 
 - (void)stopBeingDelegate {
@@ -113,17 +59,15 @@
 
   self = [super init];
   if (self) {
-    _interstitialConnector = connector;
+    _networkConnector = connector;
   }
   return self;
 }
 
 - (void)getInterstitial {
-  id<GADMAdNetworkConnector> strongConnector = _interstitialConnector;
-  NSString *gameID =
-      [[[strongConnector credentials] objectForKey:GADMAdapterUnityGameID] copy];
-  _placementID =
-      [[[strongConnector credentials] objectForKey:GADMAdapterUnityPlacementID] copy];
+  id<GADMAdNetworkConnector> strongConnector = _networkConnector;
+  NSString *gameID = [[[strongConnector credentials] objectForKey:kGADMAdapterUnityGameID] copy];
+  _placementID = [[[strongConnector credentials] objectForKey:kGADMAdapterUnityPlacementID] copy];
   if (!gameID || !_placementID) {
     NSError *error = GADUnityErrorWithDescription(@"Game ID and Placement ID cannot be nil.");
     [strongConnector adapter:self didFailAd:error];
@@ -137,7 +81,7 @@
 - (void)presentInterstitialFromRootViewController:(UIViewController *)rootViewController {
   // We will send adapterWillPresentInterstitial callback before presenting unity ad because the ad
   // has already loaded.
-  [_interstitialConnector adapterWillPresentInterstitial:self];
+  [_networkConnector adapterWillPresentInterstitial:self];
   [[GADMAdapterUnitySingleton sharedInstance]
       presentInterstitialAdForViewController:rootViewController
                                     delegate:self];
@@ -146,10 +90,17 @@
 #pragma mark Banner Methods
 
 - (void)getBannerWithSize:(GADAdSize)adSize {
-  // Unity Ads doesn't support banner ads.
-  id<GADMAdNetworkConnector> strongConnector = _interstitialConnector;
-  NSError *error = GADUnityErrorWithDescription(@"Unity Ads doesn't support banner ads.");
-  [strongConnector adapter:self didFailAd:error];
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
+  NSString *gameID =
+      [[[strongNetworkConnector credentials] objectForKey:kGADMAdapterUnityGameID] copy];
+  _placementID =
+      [[[strongNetworkConnector credentials] objectForKey:kGADMAdapterUnityPlacementID] copy];
+  if (!gameID || !_placementID) {
+    NSError *error = GADUnityErrorWithDescription(@"Game ID and Placement ID cannot be nil.");
+    [strongNetworkConnector adapter:self didFailAd:error];
+    return;
+  }
+  [[GADMAdapterUnitySingleton sharedInstance] presentBannerAd:gameID delegate:self];
 }
 
 - (BOOL)isBannerAnimationOK:(GADMBannerAnimationType)animType {
@@ -173,86 +124,98 @@
 }
 
 - (void)unityAdsDidFinish:(NSString *)placementID withFinishState:(UnityAdsFinishState)state {
-  id<GADMAdNetworkConnector> strongInterstitialConnector = _interstitialConnector;
-  id<GADMRewardBasedVideoAdNetworkConnector> strongRewardedConnector = _rewardBasedVideoAdConnector;
-  if (strongInterstitialConnector) {
-    [strongInterstitialConnector adapterWillDismissInterstitial:self];
-    [strongInterstitialConnector adapterDidDismissInterstitial:self];
-  } else if (strongRewardedConnector) {
-    if (state == kUnityAdsFinishStateCompleted) {
-      [strongRewardedConnector adapterDidCompletePlayingRewardBasedVideoAd:self];
-      // Unity Ads doesn't provide a way to set the reward on their front-end. Default to a reward
-      // amount of 1. Publishers using this adapter should override the reward on the AdMob
-      // front-end.
-      GADAdReward *reward =
-          [[GADAdReward alloc] initWithRewardType:@"" rewardAmount:[NSDecimalNumber one]];
-      [strongRewardedConnector adapter:self didRewardUserWithReward:reward];
-    }
-    [strongRewardedConnector adapterDidCloseRewardBasedVideoAd:self];
-  }
-}
-
-- (void)unityAdsDidStart:(NSString *)placementID {
-  id<GADMRewardBasedVideoAdNetworkConnector> strongRewardedConnector = _rewardBasedVideoAdConnector;
-  if (strongRewardedConnector) {
-    [strongRewardedConnector adapterDidStartPlayingRewardBasedVideoAd:self];
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapterWillDismissInterstitial:self];
+    [strongNetworkConnector adapterDidDismissInterstitial:self];
   }
 }
 
 - (void)unityAdsReady:(NSString *)placementID {
-  id<GADMAdNetworkConnector> strongInterstitialConnector = _interstitialConnector;
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
   if (!_isLoading) {
     return;
   }
 
-  if (strongInterstitialConnector) {
-    [strongInterstitialConnector adapterDidReceiveInterstitial:self];
-  } else {
-    [_rewardBasedVideoAdConnector adapterDidReceiveRewardBasedVideoAd:self];
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapterDidReceiveInterstitial:self];
   }
   _isLoading = NO;
 }
 
 - (void)unityAdsDidClick:(NSString *)placementID {
-  id<GADMAdNetworkConnector> strongInterstitialConnector = _interstitialConnector;
-  id<GADMRewardBasedVideoAdNetworkConnector> strongRewardedConnector = _rewardBasedVideoAdConnector;
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
   // The Unity Ads SDK doesn't provide an event for leaving the application, so the adapter assumes
   // that a click event indicates the user is leaving the application for a browser or deeplink, and
   // notifies the Google Mobile Ads SDK accordingly.
-  if (strongInterstitialConnector) {
-    [strongInterstitialConnector adapterDidGetAdClick:self];
-    [strongInterstitialConnector adapterWillLeaveApplication:self];
-  } else {
-    [strongRewardedConnector adapterDidGetAdClick:self];
-    [strongRewardedConnector adapterWillLeaveApplication:self];
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapterDidGetAdClick:self];
+    [strongNetworkConnector adapterWillLeaveApplication:self];
   }
 }
 
 - (void)unityAdsDidError:(UnityAdsError)error withMessage:(NSString *)message {
-  id<GADMAdNetworkConnector> strongInterstitialConnector = _interstitialConnector;
-  id<GADMRewardBasedVideoAdNetworkConnector> strongRewardedConnector = _rewardBasedVideoAdConnector;
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
   if (!_isLoading) {
     // Unity Ads show error will only happen after the ad has been loaded. So, we will send
     // dismiss/close callbacks.
     if (error == kUnityAdsErrorShowError) {
-      if (strongInterstitialConnector) {
-        [strongInterstitialConnector adapterWillDismissInterstitial:self];
-        [strongInterstitialConnector adapterDidDismissInterstitial:self];
-      } else {
-        [strongRewardedConnector adapterDidCloseRewardBasedVideoAd:self];
+      if (strongNetworkConnector) {
+        [strongNetworkConnector adapterWillDismissInterstitial:self];
+        [strongNetworkConnector adapterDidDismissInterstitial:self];
       }
     }
     return;
   }
 
   NSError *errorWithDescription = GADUnityErrorWithDescription(message);
-  if (strongInterstitialConnector) {
-    [strongInterstitialConnector adapter:self didFailAd:errorWithDescription];
-  } else if (strongRewardedConnector) {
-    [strongRewardedConnector adapter:self
-        didFailToLoadRewardBasedVideoAdwithError:errorWithDescription];
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapter:self didFailAd:errorWithDescription];
   }
   _isLoading = NO;
+}
+
+- (void)unityAdsDidStart:(nonnull NSString *)placementId {
+  // nothing to do
+}
+
+#pragma mark - Unity Banner Delegate Methods
+
+- (void)unityAdsBannerDidClick:(nonnull NSString *)placementId {
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapterDidGetAdClick:self];
+    [strongNetworkConnector adapterWillLeaveApplication:self];
+  }
+}
+
+- (void)unityAdsBannerDidError:(nonnull NSString *)message {
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
+  if (strongNetworkConnector) {
+    NSError *error = GADUnityErrorWithDescription(@"Unity Ads Banner internal error");
+    [strongNetworkConnector adapter:self didFailAd:error];
+  }
+}
+
+- (void)unityAdsBannerDidHide:(nonnull NSString *)placementId {
+  NSLog(@"Unity Ads Banner did hide.");
+}
+
+- (void)unityAdsBannerDidLoad:(nonnull NSString *)placementId view:(nonnull UIView *)view {
+  id<GADMAdNetworkConnector> strongNetworkConnector = _networkConnector;
+  if (strongNetworkConnector) {
+    [strongNetworkConnector adapter:self didReceiveAdView:view];
+  } else {
+    NSLog(@"ERROR: Network connector for UnityAds banner adapter not found.");
+  }
+}
+
+- (void)unityAdsBannerDidShow:(nonnull NSString *)placementId {
+  NSLog(@"Unity Ads Banner is showing.");
+}
+
+- (void)unityAdsBannerDidUnload:(nonnull NSString *)placementId {
+  NSLog(@"Unity Ads Banner has unloaded.");
 }
 
 @end